+++ Ilert/ilert/action_ilert_trigger_alert.py
@@ -3,7 +3,7 @@ from typing import Annotated
 from uuid import UUID
 
 # third parties
-from pydantic import BaseModel, Field, HttpUrl, StringConstraints
+from pydantic import BaseModel, Field, HttpUrl, StringConstraints, model_validator
 from requests import Response
 from sekoia_automation.action import Action
 
@@ -15,9 +15,28 @@ NonEmptyStr = Annotated[str, StringConstraints(strip_whitespace=True, min_length
 
 
 class IlertTriggerAlertArguments(BaseModel):
-    alert_uuid: UUID = Field(..., description="The Unique identifier of the alert")
+    alert_uuid: str = Field(..., description="The identifier (UUID or short id) of the alert")
     api_key: NonEmptyStr = Field(..., description="The Sekoia.io API-Key to read the alert content.")
     base_url: HttpUrl = Field(..., description="Base URL of Sekoia.io api (e.g. https://api.sekoia.io/).")
+
+    @model_validator(mode="after")
+    def validate_alert_uuid(self) -> "IlertTriggerAlertArguments":
+        # emptiness check
+        if self.alert_uuid is None or not self.alert_uuid.strip():
+            raise ValueError("The alert identifier must not be empty")
+
+        # short id validation
+        if self.alert_uuid.startswith("AL"):
+            # If the identifier starts with "AL", we assume it's a short ID
+            return self
+
+        # UUID validation
+        try:
+            UUID(self.alert_uuid)
+        except ValueError:
+            raise ValueError(f"Invalid alert identifier: {self.alert_uuid}")
+
+        return self
 
 
 class IlertTriggerAlertAction(Action):
@@ -28,12 +47,12 @@ class IlertTriggerAlertAction(Action):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-    def _get_alert(self, alert_uuid: UUID, api_key: str, base_url: HttpUrl) -> dict:
+    def _get_alert(self, alert_uuid: str, api_key: str, base_url: HttpUrl) -> dict:
         """
         Returns the definition of an alert
         """
 
-        url = f"{str(base_url).rstrip('/')}/v1/sic/alerts/{str(alert_uuid)}"
+        url = f"{str(base_url).rstrip('/')}/v1/sic/alerts/{alert_uuid}"
 
         response: Response = requests_retry_session().get(url, headers={"Authorization": f"Bearer {api_key}"})
         response.raise_for_status()