+++ Ilert/tests/test_action_ilert_triggeralert.py
@@ -120,3 +120,37 @@ def test_ilert_postalert_rejects_invalid_argument_shapes(arguments):
             action.run(arguments)
 
     assert mock.call_count == 0
+
+
+def test_ilert_postalert_accepts_short_id():
+    integration_key: str = "my-fake-integration-key"
+
+    action: IlertTriggerAlertAction = IlertTriggerAlertAction()
+    action.module.configuration = {"integration_key": integration_key}
+
+    alert_short_id = "AL98312BKWZ"
+    base_url = "https://api.sekoia.io/"
+    api_key = "XKDF84729HNQP16"
+
+    alert_info = {
+        "urgency": {"current_value": 42, "display": "medium"},
+        "short_id": alert_short_id,
+        "entity": {"name": "Red fox"},
+        "title": "Test alert for ilert",
+        "alert_type": {"category": "network", "value": "c2-traffic"},
+        "source": "172.16.0.5",
+        "target": "example.org.fake",
+        "details": "some details here",
+        "status": {"name": "Acknowledged"},
+    }
+    with requests_mock.Mocker() as mock:
+        mock.get(
+            f"{base_url}v1/sic/alerts/{alert_short_id}",
+            json=alert_info,
+        )
+        hook_url = f"{DEFAULT_EVENTS_URL}/{integration_key}"
+        mock.post(hook_url, status_code=202)
+
+        action.run({"alert_uuid": alert_short_id, "api_key": api_key, "base_url": base_url})
+
+        assert mock.call_count == 2